Start of removing database from authentication tests
